requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
paystackapi==2.0.0
platformdirs==2.5.1
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
zipp==3.7.0
